+++ Assets/Source/RosterGen/Character.cs
@@ -7,7 +7,7 @@ public class Character
 {
     //Demographics: they are physical, undeniable descriptions of a person that can be discovered simply by seeing / meeting them.
     public int rosterId; // Where in the roster list of known characters (and sprites) this person is.
-    public uint simulatedId; // The unique ID from (0 - rosterSize - 1) that contains all this character's constrainable CPD values
+    public ulong simulatedId; // The unique ID from (0 - rosterSize - 1) that contains all this character's constrainable CPD values
                      // All other (cosmetic) random values generated using this simulatedId as a seed
     Dictionary<CPD_Type, CPD_Variant> createdCharacteristics; // Once we create a character we can assign them data in here
 
@@ -30,7 +30,7 @@ public class Character
 
     // The only thing you need to create a character is their position in the roster and their simulated ID!
     // Everything else can be determined on the fly as necessary
-    public Character(int rosterId, uint simulatedId)
+    public Character(int rosterId, ulong simulatedId)
     {
         this.rosterId = rosterId;
         this.simulatedId = simulatedId;
@@ -57,20 +57,20 @@ public class Character
         // and will not be reset until we call it again.
 
         bool isMale = true;
-        uint workingSeed = simulatedId;
+        ulong workingSeed = simulatedId;
         if(createdCharacteristics.ContainsKey(CPD_Type.Gender))
         {
             int gender = createdCharacteristics[CPD_Type.Gender].cpdID;
             if (gender < 2) isMale = gender == 0;
             else
             {
-                (uint s, int v) nbName = CharRandomValue.RangedSeedRandomizer(simulatedId, 0, 2);
+                (ulong s, int v) nbName = CharRandomValue.RangedSeedRandomizer(simulatedId, 0, 2);
                 workingSeed = nbName.s;
                 isMale = nbName.v == 0;
             }
         }
         Debug.Log("Will use male name: " + isMale);
-        (uint s, string f, string l) fullName = CharRandomValue.CRV_randomName(workingSeed, isMale);
+        (ulong s, string f, string l) fullName = CharRandomValue.CRV_randomName(workingSeed, isMale);
         firstName = fullName.f;
         lastName = fullName.l;
     }
@@ -116,7 +116,7 @@ public class Character
     /// <summary>
     /// Gets the color value from a CPD assumed to be a color
     /// </summary>
-    public (uint, Color) getColorField(uint seed, CPD_Type cpdType)
+    public (ulong, Color) getColorField(ulong seed, CPD_Type cpdType)
     {
         return (createdCharacteristics[cpdType].critVal as CPD_CritVal_Color).col.getColor(simulatedId);
     }
