+++ Assets/Source/UI/CharacterCard.cs
@@ -6,7 +6,7 @@ using TMPro;
 
 public class CharacterCard : MonoBehaviour
 {
-    public uint characterId;
+    public ulong characterId;
 
     [SerializeField] SpriteRenderer portraitFrame;
     [SerializeField] SpriteRenderer portrait;
@@ -17,7 +17,7 @@ public class CharacterCard : MonoBehaviour
 
     [SerializeField] TextMeshProUGUI cityAbbrText;
 
-    public static event Action<uint> charCardClicked = (_) => { };
+    public static event Action<ulong> charCardClicked = (_) => { };
 
     // Start is called before the first frame update
     void Start()
@@ -38,14 +38,14 @@ public class CharacterCard : MonoBehaviour
         cityAbbrText.enabled = false;
         drawMat = portrait.material;
 
-        uint startingSeed = c.simulatedId;
+        ulong startingSeed = c.simulatedId;
 
         // Main portrait
-        (uint s, int v) crv1 = CharRandomValue.RangedSeedRandomizer(startingSeed, 0, 3);
+        (ulong s, int v) crv1 = CharRandomValue.RangedSeedRandomizer(startingSeed, 0, 3);
         drawMat.SetInt("_BodyIdx", crv1.v);
-        (uint s, int v) crv2 = CharRandomValue.RangedSeedRandomizer(crv1.s, 0, 8);
+        (ulong s, int v) crv2 = CharRandomValue.RangedSeedRandomizer(crv1.s, 0, 8);
         drawMat.SetInt("_HeadIdx", crv2.v);
-        (uint s, int v) crv3 = CharRandomValue.RangedSeedRandomizer(crv2.s, 0, 14);
+        (ulong s, int v) crv3 = CharRandomValue.RangedSeedRandomizer(crv2.s, 0, 14);
         drawMat.SetInt("_FaceIdx", crv3.v);
 
         int weight = c.getCategoryIndexofCharacteristic(CPD_Type.Weight);
@@ -56,16 +56,16 @@ public class CharacterCard : MonoBehaviour
 
         int Hairlen = c.getCategoryIndexofCharacteristic(CPD_Type.HairStyle);
         drawMat.SetInt("_HairLength", Hairlen);
-        (uint s, int v) crv4 = CharRandomValue.randomHairIndex(crv3.s, Hairlen);
+        (ulong s, int v) crv4 = CharRandomValue.randomHairIndex(crv3.s, Hairlen);
         drawMat.SetInt("_HairIdx", crv4.v);
 
-        (uint s, Color v) crv5 = c.getColorField(crv4.s, CPD_Type.HairColor);
+        (ulong s, Color v) crv5 = c.getColorField(crv4.s, CPD_Type.HairColor);
         drawMat.SetColor("_HairColor", crv5.v);
-        (uint s, Color v) crv6 = c.getColorField(crv5.s, CPD_Type.SkinTone);
+        (ulong s, Color v) crv6 = c.getColorField(crv5.s, CPD_Type.SkinTone);
         drawMat.SetColor("_SkinColor", crv6.v);
-        (uint s, Color v) crv7 = c.getColorField(crv6.s, CPD_Type.EyeColor);
+        (ulong s, Color v) crv7 = c.getColorField(crv6.s, CPD_Type.EyeColor);
         drawMat.SetColor("_EyeColor", crv7.v);
-        (uint s, Color v) crv8 = c.getColorField(crv7.s, CPD_Type.FavoriteColor);
+        (ulong s, Color v) crv8 = c.getColorField(crv7.s, CPD_Type.FavoriteColor);
         drawMat.SetColor("_BodyColor", crv8.v);
 
         // Locations
